+++ test/models_test.rb
@@ -1,0 +1,7 @@
+require_relative "test_helper"
+
+class ModelsTest < Minitest::Test
+  def test_resnet
+    net = TorchVision::Models::ResNet18.new
+  end
+end