+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.2 (unreleased)
 
 - Added CIFAR10, CIFAR100, FashionMNIST, and KMNIST datasets
+- Added ResNet18 model
 
 ## 0.1.1 (2020-04-28)